refactor: model, fix, view
