+++ App/Views/admin/event.php
@@ -1,0 +1,3 @@
+<div class="container">
+  <h2>Events page</h2>
+</div>
